refactor: extract SeriesService from MediaManager
Extract SeriesService as a domain-focused service handling series CRUD,
TV detail lookups, series indexing, and library listing operations.

New files:
- pkg/manager/series_service.go: SeriesService struct with constructor
  and all series-related methods extracted from MediaManager
- pkg/manager/series_service_test.go: 18 unit tests for SeriesService

Modified files:
- pkg/manager/manager.go: Added seriesService field, delegation methods
  replace inline implementations, New() creates SeriesService
- pkg/manager/series_index.go: Reduced to delegation stubs for methods
  still called by series_reconcile.go
- pkg/manager/tv_details.go: Reduced to delegation stub

Design decisions:
- SeriesMetadataProvider interface decouples SeriesService from
  MediaManagers metadata subsystem (to be extracted later)
- Reconciliation methods stay on MediaManager due to deep cross-cutting
  dependencies on indexers, downloads, quality, and search
- Package-level helper functions (buildEpisodeResult, matchEpisodeFile,
  modelEpisodeFile, filterAndMap, parseMediaResult) remain shared
- Types (TVDetailResult, SeasonResult, etc.) stay in manager_types.go

Part of #139
