+++ pkg/manager/manager.go
@@ -42,6 +42,7 @@ type MediaManager struct {
 	downloadClientService *DownloadClientService
 	qualityService        *QualityService
 	jobService            *JobService
+	seriesService         *SeriesService
 	config                config.Config
 	configs               config.Manager
 }
@@ -64,6 +65,8 @@ func New(tmbdClient tmdb.ITmdb, indexerFactory indexer.Factory, library library.
 		config:                fullConfig,
 		configs:               managerConfigs,
 	}
+
+	m.seriesService = NewSeriesService(tmbdClient, library, store, store, m.qualityService, &m)
 
 	executors := map[JobType]JobExecutor{
 		MovieReconcile: func(ctx context.Context, jobID int64) error {
@@ -174,273 +177,22 @@ func (m MediaManager) GetMovieDetailByTMDBID(ctx context.Context, tmdbID int) (*
 
 // GetTVDetailByTMDBID retrieves detailed information for a single TV show by TMDB ID
 func (m MediaManager) GetTVDetailByTMDBID(ctx context.Context, tmdbID int) (*TVDetailResult, error) {
-	log := logger.FromCtx(ctx)
-
-	// Get data from various sources
-	metadata, seriesDetailsResponse, err := m.getTVMetadataAndDetails(ctx, tmdbID)
-	if err != nil {
-		log.Error("failed to get TV metadata and details", zap.Error(err), zap.Int("tmdbID", tmdbID))
-		return nil, err
-	}
-
-	// Get library information
-	series, err := m.seriesStorage.GetSeries(ctx, table.Series.SeriesMetadataID.EQ(sqlite.Int32(metadata.ID)))
-	if err != nil && !errors.Is(err, storage.ErrNotFound) {
-		log.Debug("error checking series library status", zap.Error(err), zap.Int32("metadataID", metadata.ID))
-	}
-
-	// Get seasons and episodes information for the consolidated response
-	var seasons []SeasonResult
-	if series != nil {
-		seasonsData, err := m.getSeasonsWithEpisodes(ctx, series.ID)
-		if err != nil {
-			log.Debug("failed to get seasons and episodes", zap.Error(err), zap.Int32("metadataID", metadata.ID))
-			// Continue without seasons data - buildTVDetailResult will handle empty seasons
-			seasons = []SeasonResult{}
-		} else {
-			seasons = seasonsData
-		}
-	} else {
-		seasons = []SeasonResult{}
-	}
-
-	// Transform data into result
-	result := m.buildTVDetailResult(metadata, seriesDetailsResponse, series, seasons)
-
-	// Add stored external IDs from database
-	if extIDs, err := DeserializeExternalIDs(metadata.ExternalIds); err == nil && extIDs != nil {
-		result.ExternalIDs = &ExternalIDs{ImdbID: extIDs.ImdbID, TvdbID: extIDs.TvdbID}
-	}
-
-	// Add stored watch providers from database
-	if wpData, err := DeserializeWatchProviders(metadata.WatchProviders); err == nil && wpData != nil {
-		providers := make([]WatchProvider, 0, len(wpData.US.Flatrate))
-		for _, p := range wpData.US.Flatrate {
-			providers = append(providers, WatchProvider{
-				ProviderID: p.ProviderID,
-				Name:       p.Name,
-				LogoPath:   p.LogoPath,
-			})
-		}
-		result.WatchProviders = providers
-	}
-
-	return result, nil
+	return m.seriesService.GetTVDetailByTMDBID(ctx, tmdbID)
 }
 
-// buildTVDetailResult transforms metadata and TMDB details into TVDetailResult
+// buildTVDetailResult is delegated to SeriesService
 func (m MediaManager) buildTVDetailResult(metadata *model.SeriesMetadata, details *tmdb.SeriesDetailsResponse, series *storage.Series, seasons []SeasonResult) *TVDetailResult {
-	result := &TVDetailResult{
-		TMDBID:        metadata.TmdbID,
-		Title:         metadata.Title,
-		SeasonCount:   metadata.SeasonCount,
-		EpisodeCount:  metadata.EpisodeCount,
-		LibraryStatus: "Not In Library", // Default status
-	}
-
-	if metadata.Overview != nil {
-		result.Overview = *metadata.Overview
-	}
-
-	// Set poster path - prefer database over TMDB API to avoid unnecessary API calls
-	if metadata.PosterPath != nil && *metadata.PosterPath != "" {
-		result.PosterPath = *metadata.PosterPath
-	} else if details.PosterPath != "" {
-		result.PosterPath = details.PosterPath
-	}
-
-	// Set backdrop path from TMDB API (not stored in database)
-	if details.BackdropPath != "" {
-		result.BackdropPath = &details.BackdropPath
-	}
-
-	// Format dates
-	if metadata.FirstAirDate != nil {
-		firstAirDateStr := metadata.FirstAirDate.Format("2006-01-02")
-		result.FirstAirDate = &firstAirDateStr
-	}
-	if metadata.LastAirDate != nil {
-		lastAirDateStr := metadata.LastAirDate.Format("2006-01-02")
-		result.LastAirDate = &lastAirDateStr
-	}
-	// Next air date from TMDB next episode
-	if details.NextEpisodeToAir.AirDate != "" {
-		nextAir := details.NextEpisodeToAir.AirDate
-		result.NextAirDate = &nextAir
-	}
-
-	// Status when available
-	if details.Status != "" {
-		status := details.Status
-		result.Status = &status
-	}
-
-	// Original language
-	if details.OriginalLanguage != "" {
-		ol := details.OriginalLanguage
-		result.OriginalLanguage = &ol
-	}
-
-	// Production countries (names when available)
-	if len(details.ProductionCountries) > 0 {
-		pcs := make([]string, 0, len(details.ProductionCountries))
-		for _, pc := range details.ProductionCountries {
-			if pc.Name != nil && *pc.Name != "" {
-				pcs = append(pcs, *pc.Name)
-			} else if pc.Iso31661 != nil && *pc.Iso31661 != "" {
-				pcs = append(pcs, *pc.Iso31661)
-			}
-		}
-		result.ProductionCountries = pcs
-	}
-
-	// Networks with optional logos
-	if len(details.Networks) > 0 {
-		networks := make([]NetworkInfo, 0, len(details.Networks))
-		for _, n := range details.Networks {
-			ni := NetworkInfo{Name: n.Name}
-			if n.LogoPath != "" {
-				lp := n.LogoPath
-				ni.LogoPath = &lp
-			}
-			networks = append(networks, ni)
-		}
-		result.Networks = networks
-	}
-
-	// Extract genre names
-	if len(details.Genres) > 0 {
-		var genres []string
-		for _, genre := range details.Genres {
-			genres = append(genres, genre.Name)
-		}
-		result.Genres = genres
-	}
-
-	// Set additional fields from TMDB response
-	if details.Adult {
-		result.Adult = &details.Adult
-	}
-	if details.Popularity > 0 {
-		pop := float64(details.Popularity)
-		result.Popularity = &pop
-	}
-	// Map ratings when available
-	if details.VoteAverage > 0 {
-		va := float32(details.VoteAverage)
-		result.VoteAverage = &va
-	}
-	if details.VoteCount > 0 {
-		vc := int(details.VoteCount)
-		result.VoteCount = &vc
-	}
-
-	// Set library status information if series exists
-	if series != nil {
-		result.ID = &series.ID
-		result.LibraryStatus = string(series.State)
-		result.Path = series.Path
-		result.QualityProfileID = &series.QualityProfileID
-		monitored := series.Monitored == 1
-		result.Monitored = &monitored
-		monitorNewSeasons := series.MonitorNewSeasons == 1
-		result.MonitorNewSeasons = &monitorNewSeasons
-	}
-
-	// Add seasons information if available
-	if len(seasons) > 0 {
-		result.Seasons = seasons
-	}
-
-	return result
+	return m.seriesService.buildTVDetailResult(metadata, details, series, seasons)
 }
 
 // getSeasonsWithEpisodes retrieves seasons and their episodes for a series
 func (m MediaManager) getSeasonsWithEpisodes(ctx context.Context, seriesID int32) ([]SeasonResult, error) {
-	log := logger.FromCtx(ctx)
-
-	// Query seasons for this series
-	seasons, err := m.seriesStorage.ListSeasons(ctx,
-		table.Season.SeriesID.EQ(sqlite.Int32(seriesID)))
-	if err != nil {
-		log.Error("failed to list seasons", zap.Error(err), zap.Int32("seriesID", seriesID))
-		return nil, err
-	}
-
-	// Transform to response format with metadata and episodes
-	var results []SeasonResult
-	for _, season := range seasons {
-		// Get season metadata for rich data
-		if season.SeasonMetadataID == nil {
-			log.Debug("season has no metadata ID, skipping", zap.Int32("seasonID", season.ID))
-			continue
-		}
-
-		seasonMeta, err := m.seriesMetaStorage.GetSeasonMetadata(ctx,
-			table.SeasonMetadata.ID.EQ(sqlite.Int32(*season.SeasonMetadataID)))
-		if err != nil {
-			log.Error("failed to get season metadata", zap.Error(err), zap.Int32("seasonMetadataID", *season.SeasonMetadataID))
-			continue
-		}
-
-		// Get episodes for this season
-		episodes, err := m.getEpisodesForSeason(ctx, season.ID, season.SeriesID, seasonMeta.Number)
-		if err != nil {
-			log.Debug("failed to get episodes for season", zap.Error(err), zap.Int32("seasonID", season.ID))
-			// Continue with empty episodes array
-			episodes = []EpisodeResult{}
-		}
-
-		result := SeasonResult{
-			SeriesID:     season.SeriesID,
-			Number:       seasonMeta.Number,
-			Title:        seasonMeta.Title,
-			TMDBID:       seasonMeta.TmdbID,
-			Monitored:    season.Monitored == 1,
-			EpisodeCount: int32(len(episodes)),
-			Episodes:     episodes,
-		}
-
-		// Add optional fields
-		if seasonMeta.Overview != nil {
-			result.Overview = seasonMeta.Overview
-		}
-		if seasonMeta.AirDate != nil {
-			airDateStr := seasonMeta.AirDate.Format("2006-01-02")
-			result.AirDate = &airDateStr
-		}
-
-		results = append(results, result)
-	}
-
-	return results, nil
+	return m.seriesService.getSeasonsWithEpisodes(ctx, seriesID)
 }
 
-// preloadEpisodeMetadata fetches all episode metadata for the given IDs in a single query,
-// returning a map keyed by metadata ID.
+// preloadEpisodeMetadata fetches all episode metadata for the given IDs in a single query.
 func (m MediaManager) preloadEpisodeMetadata(ctx context.Context, episodes []*storage.Episode) map[int32]*model.EpisodeMetadata {
-	ids := make([]sqlite.Expression, 0, len(episodes))
-	for _, ep := range episodes {
-		if ep.EpisodeMetadataID != nil {
-			ids = append(ids, sqlite.Int32(*ep.EpisodeMetadataID))
-		}
-	}
-	if len(ids) == 0 {
-		return nil
-	}
-
-	metas, err := m.seriesMetaStorage.ListEpisodeMetadata(ctx, table.EpisodeMetadata.ID.IN(ids...))
-	if err != nil {
-		log := logger.FromCtx(ctx)
-		log.Error("failed to batch fetch episode metadata", zap.Error(err))
-		return nil
-	}
-
-	result := make(map[int32]*model.EpisodeMetadata, len(metas))
-	for _, meta := range metas {
-		result[meta.ID] = meta
-	}
-	return result
+	return m.seriesService.preloadEpisodeMetadata(ctx, episodes)
 }
 
 // buildEpisodeResult constructs an EpisodeResult from a storage episode and optional metadata.
@@ -481,27 +233,7 @@ func buildEpisodeResult(episode *storage.Episode, episodeMeta *model.EpisodeMeta
 
 // getEpisodesForSeason retrieves episodes for a specific season
 func (m MediaManager) getEpisodesForSeason(ctx context.Context, seasonID int32, seriesID int32, seasonNumber int32) ([]EpisodeResult, error) {
-	log := logger.FromCtx(ctx)
-
-	// Query episodes for this season
-	episodes, err := m.seriesStorage.ListEpisodes(ctx,
-		table.Episode.SeasonID.EQ(sqlite.Int32(seasonID)))
-	if err != nil {
-		log.Error("failed to list episodes", zap.Error(err), zap.Int32("seasonID", seasonID))
-		return nil, err
-	}
-
-	results := make([]EpisodeResult, 0, len(episodes))
-	metaMap := m.preloadEpisodeMetadata(ctx, episodes)
-	for _, episode := range episodes {
-		var episodeMeta *model.EpisodeMetadata
-		if episode.EpisodeMetadataID != nil {
-			episodeMeta = metaMap[*episode.EpisodeMetadataID]
-		}
-		results = append(results, buildEpisodeResult(episode, episodeMeta, seriesID, seasonNumber))
-	}
-
-	return results, nil
+	return m.seriesService.getEpisodesForSeason(ctx, seasonID, seriesID, seasonNumber)
 }
 
 // GetConfigSummary returns a readonly summary of the library configuration
@@ -545,40 +277,13 @@ func (m MediaManager) GetEntityTransitionHistory(ctx context.Context, entityType
 	return m.jobService.GetEntityTransitionHistory(ctx, entityType, entityID)
 }
 
-// SearchMovie query tmdb for tv shows
+// SearchTV queries TMDB for TV shows
 func (m MediaManager) SearchTV(ctx context.Context, query string) (*SearchMediaResponse, error) {
-	log := logger.FromCtx(ctx)
-	if query == "" {
-		log.Debug("search tv query is empty", zap.String("query", query))
-		return nil, errors.New("query is empty")
-	}
-
-	res, err := m.tmdb.SearchTv(ctx, &tmdb.SearchTvParams{Query: query})
-	if err != nil {
-		log.Error("search tv failed request", zap.Error(err))
-		return nil, err
-	}
-	defer res.Body.Close()
-
-	log.Debug("search tv response", zap.Any("status", res.Status))
-	result, err := parseMediaResult(res)
-	if err != nil {
-		log.Debug("error parsing tv show query result", zap.Error(err))
-		return nil, err
-	}
-
-	return result, nil
+	return m.seriesService.SearchTV(ctx, query)
 }
 
 func (m MediaManager) GetSeriesDetails(ctx context.Context, tmdbID int) (model.SeriesMetadata, error) {
-	var model model.SeriesMetadata
-	det, err := m.tmdb.GetSeriesDetails(ctx, tmdbID)
-	if err != nil {
-		return model, err
-	}
-
-	model, err = FromSeriesDetails(*det)
-	return model, err
+	return m.seriesService.GetSeriesDetails(ctx, tmdbID)
 }
 
 func parseMediaResult(res *http.Response) (*SearchMediaResponse, error) {
@@ -674,34 +379,7 @@ func (m MediaManager) ListIndexers(ctx context.Context) ([]IndexerResponse, erro
 }
 
 func (m MediaManager) ListShowsInLibrary(ctx context.Context) ([]LibraryShow, error) {
-	series, err := m.seriesStorage.ListSeries(ctx)
-	if err != nil {
-		return nil, err
-	}
-	shows := filterAndMap(series, func(sp *storage.Series) (LibraryShow, bool) {
-		// Skip series without metadata - they haven't been reconciled yet
-		if sp.SeriesMetadataID == nil {
-			return LibraryShow{}, false
-		}
-		ls := LibraryShow{State: string(sp.State)}
-		if sp.Path != nil {
-			ls.Path = *sp.Path
-		}
-		meta, err := m.seriesMetaStorage.GetSeriesMetadata(ctx, table.SeriesMetadata.ID.EQ(sqlite.Int(int64(*sp.SeriesMetadataID))))
-		if err != nil || meta == nil {
-			return LibraryShow{}, false
-		}
-		ls.TMDBID = meta.TmdbID
-		ls.Title = meta.Title
-		if meta.PosterPath != nil {
-			ls.PosterPath = *meta.PosterPath
-		}
-		if meta.FirstAirDate != nil {
-			ls.Year = int32(meta.FirstAirDate.Year())
-		}
-		return ls, true
-	})
-	return shows, nil
+	return m.seriesService.ListShowsInLibrary(ctx)
 }
 
 // ListMoviesInLibrary returns library movies enriched with metadata
@@ -915,125 +593,7 @@ func (m MediaManager) AddMovieToLibrary(ctx context.Context, request AddMovieReq
 
 // AddSeriesToLibrary adds a series to be managed by mediaz
 func (m MediaManager) AddSeriesToLibrary(ctx context.Context, request AddSeriesRequest) (*storage.Series, error) {
-	log := logger.FromCtx(ctx)
-
-	qualityProfile, err := m.GetQualityProfile(ctx, int64(request.QualityProfileID))
-	if err != nil {
-		log.Debug("failed to get quality profile", zap.Int32("id", request.QualityProfileID), zap.Error(err))
-		return nil, err
-	}
-
-	seriesMetadata, err := m.GetSeriesMetadata(ctx, request.TMDBID)
-	if err != nil {
-		log.Debug("failed to get series metadata", zap.Error(err))
-		return nil, err
-	}
-
-	series, err := m.seriesStorage.GetSeries(ctx, table.Series.ID.EQ(sqlite.Int32(seriesMetadata.ID)))
-	// if we find the series we dont need to add it
-	if err == nil {
-		return series, err
-	}
-	if !errors.Is(err, storage.ErrNotFound) {
-		log.Warn("couldn't find series by metadata", zap.Int32("meta_id", seriesMetadata.ID), zap.Error(err))
-		return nil, err
-	}
-
-	monitorNewSeasons := int32(0)
-	if request.MonitorNewSeasons {
-		monitorNewSeasons = 1
-	}
-	series = &storage.Series{
-		Series: model.Series{
-			SeriesMetadataID:  &seriesMetadata.ID,
-			QualityProfileID:  qualityProfile.ID,
-			Monitored:         1,
-			Path:              &seriesMetadata.Title,
-			MonitorNewSeasons: monitorNewSeasons,
-		},
-	}
-
-	state := initialSeriesState(seriesMetadata.FirstAirDate)
-
-	seriesID, err := m.seriesStorage.CreateSeries(ctx, *series, state)
-	if err != nil {
-		log.Error("failed to create new missing series", zap.Error(err))
-		return nil, err
-	}
-
-	log.Debug("created new missing series", zap.Any("series", series))
-
-	// Get series to access its metadata ID
-	seriesEntity, err := m.seriesStorage.GetSeries(ctx, table.Series.ID.EQ(sqlite.Int(seriesID)))
-	if err != nil || seriesEntity.SeriesMetadataID == nil {
-		log.Error("failed to get series or series has no metadata")
-		return nil, fmt.Errorf("series has no metadata")
-	}
-
-	where := table.SeasonMetadata.SeriesMetadataID.EQ(sqlite.Int32(*seriesEntity.SeriesMetadataID))
-	seasonMetadata, err := m.seriesMetaStorage.ListSeasonMetadata(ctx, where)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, s := range seasonMetadata {
-		season := storage.Season{
-			Season: model.Season{
-				SeriesID:         int32(seriesID),
-				SeasonMetadataID: ptr.To(s.ID),
-				Monitored:        1,
-			},
-		}
-
-		seasonID, err := m.seriesStorage.CreateSeason(ctx, season, storage.SeasonStateMissing)
-		if err != nil {
-			log.Error("failed to create season", zap.Error(err))
-			return nil, err
-		}
-
-		log.Debug("created new missing season", zap.Any("season", season))
-
-		// Get the season to access its metadata ID for proper episode metadata querying
-		seasonEntity, err := m.seriesStorage.GetSeason(ctx, table.Season.ID.EQ(sqlite.Int64(seasonID)))
-		if err != nil || seasonEntity.SeasonMetadataID == nil {
-			log.Error("failed to get season or season has no metadata linked")
-			return nil, fmt.Errorf("season has no metadata")
-		}
-
-		where := table.EpisodeMetadata.SeasonMetadataID.EQ(sqlite.Int32(*seasonEntity.SeasonMetadataID))
-
-		episodesMetadata, err := m.seriesMetaStorage.ListEpisodeMetadata(ctx, where)
-		if err != nil {
-			log.Error("failed to list episode metadata", zap.Error(err))
-			return nil, err
-		}
-
-		for _, e := range episodesMetadata {
-			episode := storage.Episode{
-				Episode: model.Episode{
-					EpisodeMetadataID: ptr.To(e.ID),
-					SeasonID:          int32(seasonID),
-					Monitored:         1,
-					EpisodeNumber:     e.Number,
-				},
-			}
-
-			_, err := m.seriesStorage.CreateEpisode(ctx, episode, storage.EpisodeStateMissing)
-			if err != nil {
-				log.Error("failed to create episode", zap.Error(err))
-				return nil, err
-			}
-
-			log.Debug("created new missing episode", zap.Any("episode", episode))
-		}
-	}
-
-	series, err = m.seriesStorage.GetSeries(ctx, table.Series.ID.EQ(sqlite.Int64(seriesID)))
-	if err != nil {
-		log.Warn("failed to get created series", zap.Error(err))
-	}
-
-	return series, err
+	return m.seriesService.AddSeriesToLibrary(ctx, request)
 }
 
 func (m MediaManager) DeleteMovie(ctx context.Context, movieID int64, deleteFiles bool) error {
@@ -1063,50 +623,7 @@ func (m MediaManager) DeleteMovie(ctx context.Context, movieID int64, deleteFile
 }
 
 func (m MediaManager) DeleteSeries(ctx context.Context, seriesID int64, deleteDirectory bool) error {
-	log := logger.FromCtx(ctx)
-
-	series, err := m.seriesStorage.GetSeries(ctx, table.Series.ID.EQ(sqlite.Int64(seriesID)))
-	if err != nil {
-		return fmt.Errorf("failed to get series: %w", err)
-	}
-
-	seasons, err := m.seriesStorage.ListSeasons(ctx, table.Season.SeriesID.EQ(sqlite.Int64(seriesID)))
-	if err != nil {
-		log.Warn("failed to get seasons for cleanup", zap.Error(err))
-	}
-
-	for _, season := range seasons {
-		episodes, err := m.seriesStorage.ListEpisodes(ctx, table.Episode.SeasonID.EQ(sqlite.Int64(int64(season.ID))))
-		if err != nil {
-			log.Warn("failed to get episodes for cleanup", zap.Int32("season_id", season.ID), zap.Error(err))
-			continue
-		}
-		for _, episode := range episodes {
-			if episode.EpisodeFileID == nil {
-				continue
-			}
-			if err := m.seriesStorage.DeleteEpisodeFile(ctx, int64(*episode.EpisodeFileID)); err != nil {
-				log.Warn("failed to delete episode file", zap.Int32("episode_file_id", *episode.EpisodeFileID), zap.Error(err))
-			}
-		}
-	}
-
-	if deleteDirectory {
-		if series.Path == nil {
-			return fmt.Errorf("cannot delete directory: series path is nil")
-		}
-
-		if err := m.library.DeleteSeriesDirectory(ctx, *series.Path); err != nil {
-			return fmt.Errorf("failed to delete series directory %s: %w", *series.Path, err)
-		}
-	}
-
-	if err := m.seriesStorage.DeleteSeries(ctx, seriesID); err != nil {
-		return fmt.Errorf("failed to delete series: %w", err)
-	}
-
-	log.Info("deleted series", zap.Int64("id", seriesID), zap.Bool("directory_deleted", deleteDirectory))
-	return nil
+	return m.seriesService.DeleteSeries(ctx, seriesID, deleteDirectory)
 }
 
 func (m MediaManager) UpdateMovieMonitored(ctx context.Context, movieID int64, monitored bool) (*storage.Movie, error) {
@@ -1146,24 +663,7 @@ func (m MediaManager) UpdateMovieQualityProfile(ctx context.Context, movieID int
 }
 
 func (m MediaManager) UpdateSeriesMonitored(ctx context.Context, seriesID int64, monitored bool) (*storage.Series, error) {
-	monitoredInt := int32(0)
-	if monitored {
-		monitoredInt = 1
-	}
-
-	seriesUpdate := model.Series{Monitored: monitoredInt}
-	err := m.seriesStorage.UpdateSeries(ctx, seriesUpdate, table.Series.ID.EQ(sqlite.Int64(seriesID)))
-	if err != nil {
-		return nil, err
-	}
-
-	series, err := m.seriesStorage.GetSeries(ctx, table.Series.ID.EQ(sqlite.Int64(seriesID)))
-	if err != nil {
-		return nil, err
-	}
-
-	logger.FromCtx(ctx).Info("updated monitoring", zap.Int64("series_id", seriesID), zap.Bool("monitored", monitored))
-	return series, nil
+	return m.seriesService.UpdateSeriesMonitored(ctx, seriesID, monitored)
 }
 
 func (m MediaManager) AddIndexer(ctx context.Context, request AddIndexerRequest) (IndexerResponse, error) {
@@ -1354,114 +854,7 @@ func (m MediaManager) ListSeasonsForSeries(ctx context.Context, tmdbID int) ([]S
 
 // ListEpisodesForSeason retrieves all episodes for a season by TMDB ID and season number
 func (m MediaManager) ListEpisodesForSeason(ctx context.Context, tmdbID int, seasonNumber int) ([]EpisodeResult, error) {
-	log := logger.FromCtx(ctx)
-
-	// Ensure series metadata exists
-	seriesMetadata, err := m.GetSeriesMetadata(ctx, tmdbID)
-	if err != nil {
-		log.Error("failed to get series metadata", zap.Error(err), zap.Int("tmdbID", tmdbID))
-		return nil, err
-	}
-
-	// Find the series record that uses this metadata
-	series, err := m.seriesStorage.GetSeries(ctx,
-		table.Series.SeriesMetadataID.EQ(sqlite.Int32(seriesMetadata.ID)))
-	if err != nil {
-		log.Error("failed to get series", zap.Error(err), zap.Int32("metadataID", seriesMetadata.ID))
-		return nil, err
-	}
-
-	// Find all seasons for this series
-	seasons, err := m.seriesStorage.ListSeasons(ctx,
-		table.Season.SeriesID.EQ(sqlite.Int32(series.ID)))
-	if err != nil {
-		log.Error("failed to list seasons", zap.Error(err), zap.Int32("seriesID", series.ID))
-		return nil, err
-	}
-
-	// Find the season that matches the requested season number
-	// Priority: 1) Season with metadata matching number, 2) Season with matching season_number
-	var candidateSeasons []*storage.Season
-	var candidateMetas []*model.SeasonMetadata
-
-	for _, season := range seasons {
-		// Check if season has metadata with the right number
-		if season.SeasonMetadataID != nil {
-			meta, err := m.seriesMetaStorage.GetSeasonMetadata(ctx,
-				table.SeasonMetadata.ID.EQ(sqlite.Int32(*season.SeasonMetadataID)))
-			if err == nil && meta.Number == int32(seasonNumber) {
-				candidateSeasons = append(candidateSeasons, season)
-				candidateMetas = append(candidateMetas, meta)
-			}
-		}
-	}
-
-	// If no seasons found with metadata, fall back to season_number matching
-	if len(candidateSeasons) == 0 {
-		for _, season := range seasons {
-			if season.SeasonNumber == int32(seasonNumber) {
-				candidateSeasons = append(candidateSeasons, season)
-				candidateMetas = append(candidateMetas, nil)
-			}
-		}
-	}
-
-	if len(candidateSeasons) == 0 {
-		log.Error("season not found", zap.Int32("seriesID", seriesMetadata.ID), zap.Int("seasonNumber", seasonNumber))
-		return nil, fmt.Errorf("season %d not found for series %d", seasonNumber, tmdbID)
-	}
-
-	// If we have multiple candidates, prefer the one with episodes
-	var targetSeason *storage.Season
-	var seasonMeta *model.SeasonMetadata
-
-	if len(candidateSeasons) == 1 {
-		targetSeason = candidateSeasons[0]
-		seasonMeta = candidateMetas[0]
-	} else {
-		// Multiple candidates - check which one has episodes
-		for i, season := range candidateSeasons {
-			episodes, err := m.seriesStorage.ListEpisodes(ctx,
-				table.Episode.SeasonID.EQ(sqlite.Int32(season.ID)))
-			if err == nil && len(episodes) > 0 {
-				targetSeason = season
-				seasonMeta = candidateMetas[i]
-				break
-			}
-		}
-
-		// If no season has episodes, just use the first one
-		if targetSeason == nil {
-			targetSeason = candidateSeasons[0]
-			seasonMeta = candidateMetas[0]
-		}
-	}
-
-	// Query episodes for this season
-	episodes, err := m.seriesStorage.ListEpisodes(ctx,
-		table.Episode.SeasonID.EQ(sqlite.Int32(targetSeason.ID)))
-	if err != nil {
-		log.Error("failed to list episodes", zap.Error(err), zap.Int32("seasonID", targetSeason.ID))
-		return nil, err
-	}
-
-	// Determine season number for response
-	seasonNum := int32(seasonNumber)
-	if seasonMeta != nil {
-		seasonNum = seasonMeta.Number
-	}
-
-	results := make([]EpisodeResult, 0, len(episodes))
-	metaMap := m.preloadEpisodeMetadata(ctx, episodes)
-	for _, episode := range episodes {
-		var episodeMeta *model.EpisodeMetadata
-		if episode.EpisodeMetadataID != nil {
-			episodeMeta = metaMap[*episode.EpisodeMetadataID]
-		}
-		results = append(results, buildEpisodeResult(episode, episodeMeta, series.ID, seasonNum))
-	}
-
-	return results, nil
+	return m.seriesService.ListEpisodesForSeason(ctx, tmdbID, seasonNumber)
 }
 
 func (m MediaManager) CreateJob(ctx context.Context, request TriggerJobRequest) (JobResponse, error) {
